Safe upgrade point
I went through and had gotten everything more updated, but ran into a stupid conflict near the end that made me have to undo everything. Things should be good here even if there are major updates we are behind for now.

The biggest thing is that our recurring dates plugin requires react 18. Once we have that we can do the upgrade, but that comes with introducing 8 high vulnerabilities, although 5 of them come from sanity/vision which we could ignore because the input from validated users should be safe